requirements.txt
asgiref==3.5.2
certifi==2020.6.20
dj-database-url==0.5.0
Django==3.2.15
django-extensions==3.2.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
mccabe==0.6.1
pip==21.2.2
psycopg2==2.8.6
pyflakes==2.3.1
pytz==2022.2.1
setuptools==58.0.4
sqlparse==0.4.2
typing_extensions==4.1.1
wheel==0.37.1
whitenoise==6.2.0
zipp==3.6.0
